requirements.txt
# Core AI Orchestration
langchain
langchain-community
langchain-core

# Document Processing
pypdf
tiktoken
python-dotenv

# Vector Database (The Memory)
chromadb
langchain-chroma
langchain-huggingface
sentence-transformers

# Brains (LLM Providers)
langchain-google-genai
langchain-ollama

# Professional UI
streamlit